requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
billiard==4.2.1
boto3==1.35.78
botocore==1.35.78
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==44.0.0
defusedxml==0.8.0rc2
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.2
django-allauth==65.3.0
django-colorfield==0.11.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-grappelli==4.0.1
django-import-export==4.2.1
django-json-widget==2.0.1
django-redis==5.4.0
django-simple-history==3.7.0
django-storages==1.14.4
django-webpush==0.3.6
djangorestframework==3.15.2
docutils==0.21.2
flower==2.0.1
frozenlist==1.5.0
gunicorn==23.0.0
http_ece==1.2.1
humanize==4.11.0
idna==3.10
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
multidict==6.1.0
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
py-vapid==1.9.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-pptx==1.0.2
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
pywebpush==2.0.3
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
suit==2.0.2
tablib==3.7.0
text-unidecode==1.3
tornado==6.4.2
twilio==9.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
XlsxWriter==3.2.0
yarl==1.17.1
